[chore] 시간 포맷 변환 함수 파일 위치 이동
